refactor(proposal): Rename for review proposal request.
